unet modification. first trained model
Unet modification: When trained with several datasets, in folder trained_models another folder for the model is generated, and for every dataset another folder under this one is created with the best_model.pth and the json with the learning metrics. Now, the final best model should also be copied at the parent folder.